+++ package.json
@@ -16,6 +16,8 @@
     "changelog": "conventional-changelog -p angular -i CHANGELOG.md -s"
   },
   "dependencies": {
+    "@ai-sdk/openai": "^1.1.15",
+    "@ai-sdk/react": "^1.1.19",
     "@anthropic-ai/sdk": "^0.32.1",
     "@emoji-mart/data": "^1.2.1",
     "@emoji-mart/react": "^1.1.1",
@@ -31,7 +33,7 @@
     "@radix-ui/react-toggle": "^1.1.2",
     "@radix-ui/react-toggle-group": "^1.1.2",
     "@radix-ui/react-tooltip": "^1.1.8",
-   
+    "ai": "^4.1.47",
     "change-case": "^5.4.4",
     "class-variance-authority": "^0.7.1",
     "classnames": "^2.5.1",
@@ -41,6 +43,7 @@
     "dexie-react-hooks": "^1.1.7",
     "electron-squirrel-startup": "^1.0.1",
     "emoji-mart": "^5.6.0",
+    "express": "^4.21.2",
     "formik": "^2.4.6",
     "highlight.js": "^11.10.0",
     "immer": "^10.1.1",
@@ -66,10 +69,10 @@
     "uuid": "^10.0.0",
     "wavesurfer.js": "^7.8.6",
     "whatwg-fetch": "^3.6.20",
-    "yancey-js-util": "^3.2.0"
+    "yancey-js-util": "^3.2.0",
+    "zod": "^3.24.2"
   },
   "devDependencies": {
-    "@tailwindcss/vite": "^4.0.9",
     "@commitlint/cli": "^19.5.0",
     "@commitlint/config-conventional": "^19.5.0",
     "@electron-forge/cli": "^7.6.0",
@@ -82,7 +85,9 @@
     "@electron-forge/plugin-vite": "^7.6.0",
     "@electron/fuses": "^1.8.0",
     "@playwright/test": "^1.50.0",
+    "@tailwindcss/vite": "^4.0.9",
     "@types/electron-squirrel-startup": "^1.0.2",
+    "@types/express": "^5.0.0",
     "@types/luxon": "^3.4.2",
     "@types/node": "^22.7.5",
     "@types/react": "^18.3.11",
