+++ src/renderer.tsx
@@ -26,11 +26,34 @@
  * ```
  */
 
-import { Provider } from 'jotai'
-import { StrictMode } from 'react'
+import { Provider, useAtomValue } from 'jotai'
+import { FC, StrictMode } from 'react'
 import ReactDOM from 'react-dom/client'
+import Conversation from 'src/components/Conversation'
+import Loading from 'src/components/Loading'
+import Settings from 'src/components/Settings'
+import Sidebar from 'src/components/Sidebar'
+import { useInitial, useOnline } from 'src/hooks'
+import { configurationAtom } from 'src/stores/conversation'
+import { settingsAtom } from 'src/stores/global'
 import './assets/styles/index.css'
-import Layouts from './layouts'
+
+const Layouts: FC = () => {
+  const configuration = useAtomValue(configurationAtom)
+  const settings = useAtomValue(settingsAtom)
+  useOnline()
+  useInitial()
+
+  if (!configuration || !settings) return <Loading />
+
+  return (
+    <section className="flex">
+      <Sidebar />
+      <Conversation />
+      <Settings />
+    </section>
+  )
+}
 
 ReactDOM.createRoot(document.getElementById('root') as HTMLElement).render(
   <StrictMode>
